eurobot_ws/src/gui_zdc/requirements.txt
numpy==2.1.2
opencv-python==4.10.0.84
pillow==10.4.0
psutil==6.1.1
ping3
RPi.GPIO

